User Story C3 R1
- ADDED setting view (+ all required MVVM components)
- MODIFIED moved working hours and password changing functionality from dashboard to the settings view
