+++ Client/src/viewmodel/DashboardViewModel.java
@@ -111,39 +111,11 @@ public class DashboardViewModel implements DashBoardViewModelInterface {
     }
 
     @Override
-    public void setOpeningHours(String s) throws RuntimeException {
-        model.setOpeningHours(s);
-    }
-
-    @Override
-    public void setClosingHours(String s) throws RuntimeException {
-        model.setClosingHours(s);
-    }
-
-    @Override
-    public String getOpeningHours() {
-        return model.getOpeningHours();
-    }
-
-    @Override
-    public String getClosingHours() {
-        return model.getClosingHours();
-    }
-
-    @Override
     public void startUpdateThread() {
         this.updateThread.start();
     }
 
-    @Override
-    public boolean masterCheck(String s) {
-        return model.masterCheck(s);
-    }
 
-    @Override
-    public void updatePassword(String role, String password) {
-        model.updatePassword(role, password);
-    }
 
     public SimpleStringProperty getDateProperty() {
         return dateProperty;
