+++ Client/src/viewmodel/ViewModelFactory.java
@@ -4,10 +4,13 @@ public class ViewModelFactory {
     private final StartViewModel startViewModel;
     private final DashboardViewModel dashboardViewModel;
 
+    private final SettingsViewModel settingsViewModel;
+
     public ViewModelFactory(Model model) {
         ViewModelState viewModelState = new ViewModelState();
         startViewModel = new StartViewModel(model, viewModelState);
         dashboardViewModel = new DashboardViewModel(model, viewModelState);
+        settingsViewModel = new SettingsViewModel(model, viewModelState);
     }
 
     public StartViewModel getStartViewModel() {
@@ -16,5 +19,9 @@ public class ViewModelFactory {
 
     public DashboardViewModel getDashboardViewModel() {
         return dashboardViewModel;
+    }
+
+    public SettingsViewModel getSettingsViewModel() {
+        return settingsViewModel;
     }
 }
